requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
coverage==7.6.12
Django==3.0.11
djangorestframework==3.11.2
drf-yasg==1.21.8
flake8==7.1.1
inflection==0.5.1
mccabe==0.7.0
mysqlclient==2.2.7
packaging==24.2
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
